requirements.txt
requests
python-dotenv
supabase
psycopg2-binary
